Muting and transmission actions revision
Initially, when connecting to a server, voice transmission  and mute
are always off, but their state is preserved when going background,
screen is locked or another activity is started. Mute button affects
only the general state not disturbing automatic mute on transmission
if it is enabled in preferences. In that case, being pressed when
transmission is on, it does not unmute immediately, but instead, has
an influence upon the state after transmission. Automatic mute on
transmission now works in voice activation mode as well.