Using dummy channels to publish info
When maximum users in a channel is set to 0 its item is displayed as a
pure info with multiline text fields and without button. These
channels are not joinable anyway, and it appears to be more or less
common practice to use such items for publishing various announcements
and info messages.