Sound mode management fix in Android client
Correct audio stream routing management in Android appears to be quite
tricky when bluetooth and wired headsets are taken in account. But now
it seems to be ok in general. The only drawback is that if headset is
disconnected then active server connection should be re-established to
restore proper audio mode.