Replace jlong TTInstance with mapping to Object.hashCode()
TeamTalkBase.initTeamTalkPoll() returns a jlong that is really a
casted TTInstance. With this change the mapping from "jobject thiz" to
TTInstance is done using a std::map where the key is
Object.hashCode() and the value is TTInstance*

This simplifies dk.bearware.TeamTalkBase class since "double" native
method are no longer needed.