Fix crash issue when toggling echo cancellation off
When duplex mode is disabled (due to echo cancellation = off) then
WebRTC preprocessor is still active in echo cancellation mode. WebRTC
then tries to access an AudioFrame::output_samples which is no longer
available.

In other words the WebRTC preprocessor is configured in a way that is
incompatible with the sound device configuration. WebRTC can only echo
cancel when sound devices are initialized with
TT_InitSoundDuplexDevices().